refactor(EE10.04): extract format.ts from orchestrator.ts
Move all string rendering functions into tools/delivery/format.ts.
Migrate formatAdvanceBoundaryGuidance and formatCurrentTicketStatus
describe blocks verbatim into test/format.test.ts. Also moves
resolveEffectiveAdvanceBoundaryMode and RepairStateResult type.
orchestrator.ts re-exports all format surface unchanged.
